requirements.txt
numpy==1.23.5
torch==2.1.2
torchvision==0.16.2
tqdm==4.64.1
typing_extensions==4.12.2
yacs==0.1.8
einops
monai==1.3.2
Django~=5.1.1
Pillow~=9.5.0
djangorestframework~=3.14.0